pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pynrpf"
version = "0.2.0"
description = "PyNRPF implementation package for reverse power flow detection and correction."
readme = "README.MD"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "M. Syahman Samhan", email = "m.samhan@unsw.edu.au" }
]
dependencies = [
  "pyyaml>=6,<7",
  "pandas>=2.1,<3",
  "numpy>=1.26,<3",
  "scikit-learn>=1.4,<2",
  "xgboost>=2,<3",
  "holidays>=0.40,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "ruff>=0.9,<1",
  "build>=1.2,<2",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["pynrpf*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["src/pynrpf/_legacy"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
